+++ src/test/java/ru/yandex/practicum/filmorate/utils/IdentifierTest.java
@@ -8,7 +8,7 @@ class IdentifierTest {
     @Test
     public void next_ReturnAddedIs1() {
         Identifier identifier = new Identifier();
-        int nextId = identifier.next();
+        long nextId = identifier.next();
         assertEquals(1, nextId, "Сгенерированный ID должен быть 1");
     }
 
@@ -16,7 +16,7 @@ class IdentifierTest {
     public void setMaxId_ReturnRestoredId4_IfSetMaxIsId3() {
         Identifier identifier = new Identifier();
         identifier.setStartId(3);
-        int nextId = identifier.next();
+        long nextId = identifier.next();
         assertEquals(4, nextId, "Сгенерированный ID должен быть 4");
     }
